refactor: move mocking data logic to infra layer
